Update tag in dev and stage
Using VERSION and BUILD_NUMBER as the tag in dev and stage.

Updated the template to use a variable for the IMAGESTREAM_TAG since
that will change if in dev/stage or prod.

Changed DeploymentConfig trigger configuration automatic from true to false so
that a new ImageStreamTag will not trigger a redeployment.